refactor(web): split index.html into app.css/i18n.js/app.js
The single-file UI (8,175 lines: inline CSS + 23-locale i18n data +
app logic) was the biggest standing tech-debt item — hardest place to
find a label or resolve a merge conflict. index.html is now markup
only, loading app.css/i18n.js/app.js via relative <link>/<script src>
(pywebview loads it from a file:// path, so this works with no CORS
issue). scripts/check_i18n.py and scripts/gen_app_manifest.py, which
both parsed the old inline blocks out of index.html via node, are
repointed at the new files; tests/test_ui_localization.py and
tests/test_soundcheck.py now concatenate the split files before their
string-search assertions. No build-script change needed since
build_official.py already bundles the whole app/web/ directory.

chore(quality): add a curated ruff config, wire up pyright, fix findings

pyproject.toml adds a [tool.ruff] section (the project ran on ruff's
bare defaults with no config file) — a curated rule set chosen to
catch real bugs (bugbear, comprehensions, pyupgrade, isort, asyncio
misuse) while deliberately skipping line-length and ambiguous-unicode
checks that are pure noise on this comment-dense, 23-locale codebase.
BLE001 (blind-except, 228 hits) and flake8-bandit are left for a
dedicated follow-up rather than force-fit here.

pyright is new to the project (previously no type checker at all).
Ran in basic mode over app/ + scripts/ (tests/ excluded — its
monkeypatch-heavy fixtures throw a different, lower-value class of
error) and fixed all 125 findings: mostly deferred-init Optional
attributes given real declared types, plus a handful of genuine bugs
caught along the way — a stale 8-tuple unpack in
scripts/bench/run_session.py that would have crashed against
get_session_key()'s current 9-tuple return, a code path in
pipeline.py's audio-retry loop that could `raise None`, two
un-null-checked pycaw GetSpeakers()/GetMicrophone() calls, and two
deprecated pywebview/Pillow API calls that were emitting runtime
warnings. Both checks are now wired into .github/workflows/quality.yml
alongside the existing pytest/ruff steps.

+++ app/mix_core.py
@@ -53,7 +53,7 @@ class LookaheadLimiter:
     def __init__(self, fs: float, lookahead_ms: float = 1.5,
                  ceiling: float = 0.97, release_ms: float = 80.0):
         self.fs = float(fs)
-        self.L = max(1, int(round(fs * lookahead_ms / 1000.0)))
+        self.L = max(1, round(fs * lookahead_ms / 1000.0))
         self.ceil = float(ceiling)
         self.rel = float(np.exp(-1.0 / max(1.0, fs * release_ms / 1000.0)))
         self._delay = None
@@ -83,7 +83,7 @@ class LookaheadLimiter:
         # 1) Required per-sample gain (<= 1); peak across channels. nan_to_num
         # keeps a non-finite input sample from driving gr to nan/inf.
         peak = np.max(np.abs(x), axis=1) + 1e-9
-        peak = np.nan_to_num(peak, nan=1.0, posinf=np.float32(3.0e38), neginf=1e-9)
+        peak = np.nan_to_num(peak, nan=1.0, posinf=float(np.float32(3.0e38)), neginf=1e-9)
         gr = np.minimum(1.0, self.ceil / peak).astype(np.float32)
 
         # 2) Look-ahead: rolling minimum over the next L samples.
@@ -199,7 +199,7 @@ class DelayLine:
         self.max_delay = None if max_delay is None else float(max_delay)
         # Equal-power cross-fade length applied on a resync snap so the read
         # pointer's discontinuity does not click.
-        self._xfade = max(1, int(round(fs * xfade_ms / 1000.0)))
+        self._xfade = max(1, round(fs * xfade_ms / 1000.0))
         self._snapped_from: float | None = None
 
     def set_target(self, delay_samples: float) -> None:
